updated the customer java files so that inputs are now stored in database

+++ FinalProject/FinalProject/src/main/java/com/webdevuoit3230/FinalProject/repository/CustomerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.webdevuoit3230.FinalProject.model.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
-	
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+    
 }
